+++ docs/.vitepress/theme/Layout.vue
@@ -1,9 +1,14 @@
 <script setup lang="ts">
 import DefaultTheme from 'vitepress/theme';
+import A2wApiSidebarBrand from './A2wApiSidebarBrand.vue';
 
 const { Layout } = DefaultTheme;
 </script>
 
 <template>
-  <Layout />
+  <Layout>
+    <template #sidebar-nav-before>
+      <A2wApiSidebarBrand />
+    </template>
+  </Layout>
 </template>